[MERGE]: resolve conflict

+++ app/src/main/java/com/indipage/data/repositoryimpl/ArticleDetailRepositoryImpl.kt
@@ -22,7 +22,7 @@ class ArticleDetailRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun postTicketReceive(spaceId: Int): Result<Int> {
+    override suspend fun postTicketReceive(spaceId: Long): Result<Int> {
         return runCatching {
             dataSource.postTicketReceive(spaceId).code
         }
