+++ app/src/main/java/com/indipage/presentation/article/ArticleDetailViewModel.kt
@@ -50,7 +50,7 @@ class ArticleDetailViewModel @Inject constructor(
             .onFailure { Timber.d(it.message.toString()) }
     }
 
-    fun postTicketReceive(spaceId: Int) =
+    fun postTicketReceive(spaceId: Long) =
         viewModelScope.launch {
             apiRepository.postTicketReceive(spaceId)
                 .onSuccess {
